--- app/fs/files/notes.json
@@ -1,7 +1,0 @@
-[
-  {
-    "id": "37c9d7a1c59a2d064790e166",
-    "text": "mi primera nota",
-    "date": "2024-03-13T18:49:07.490Z"
-  }
-]
